ui: change font size and icons in AddContactScreen

+++ frontend/app/src/main/java/com/fiveis/xend/data/model/AddContactRequest.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class AddContactRequest(
-    @SerialName("group") val groupId: Long = 0,
+    @SerialName("group") val groupId: Int = 0,
     val name: String,
     val email: String,
     val context: AddContactRequestContext? = null
