+++ frontend/app/src/main/java/com/fiveis/xend/data/model/ContactResponse.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ContactResponse(
-    val id: Long,
-    @SerialName("group") val groupId: Long? = null,
+    val id: Int,
+    @SerialName("group") val groupId: Int? = null,
     val name: String,
     val email: String,
     val context: ContactResponseContext? = null,
@@ -16,7 +16,7 @@ data class ContactResponse(
 
 @Serializable
 data class ContactResponseContext(
-    val id: Long,
+    val id: Int,
     @SerialName("relationship_role") val relationshipRole: String? = null,
     @SerialName("relationship_details") val relationshipDetails: String? = null,
     @SerialName("personal_prompt") val personalPrompt: String? = null,
